requirements.txt
deta==1.1.0
keras==2.10.0
matplotlib==3.6.2
numpy==1.23.4
opencv_python==4.6.0.66
pandas==1.5.1
Pillow==9.3.0
scikit_image==0.19.3
skimage==0.0
streamlit==1.14.0
streamlit_image_comparison==0.0.2
tensorflow==2.10.0
